refactor

+++ build.sh
@@ -1,67 +1,94 @@
 #!/bin/bash
 
-d=false
+debug=False
+modOnly=False
+pythonOnly=False
 
-while getopts "v:d" flag
+while getopts "v:dmp" flag
 do
-    case "${flag}" in
-        v) v=${OPTARG};;
-        d) d=true;;
-    esac
+  case "${flag}" in
+    v) v=${OPTARG};;
+    d) debug=True;;
+    m) modOnly=True;;
+    p) pythonOnly=True;;
+  esac
 done
+
+
+
+echo "______PLAN:______"
+echo "Version: $v"
+echo "Debug: $debug"
+
+if [ "$pythonOnly" = true ]; then
+  echo "Python Only (without cef-server and as3)"
+elif [ "$modOnly" = true ]; then
+  echo "Mod Only (python + as3)"
+else
+  echo "Full Build"
+fi
+
+echo "_________________"
+echo ""
+
+
+folder="wotstat.cef_$v.wotmod"
+rm -rf $folder
 
 
 rm -rf ./build
 mkdir ./build
-cp -r ./res ./build
 
-# Set version
+
+# BUILD PYTHON
+echo "Build Python"
+cp -r ./mod/res ./build
+
+# Set version and debug mode
 configPath="./build/res/scripts/client/gui/mods/wotstat_cef/common/Config.py"
+mainPath="./build/res/scripts/client/gui/mods/wotstat_cef/WotstatWidget.py"
 perl -i -pe "s/{{VERSION}}/$v/g" "$configPath"
-
-# Set debug mode
-utilsPath="./build/res/scripts/client/gui/mods/wotstat_cef/__init__.py"
-if [ "$d" = true ]; then
-    echo "Building DEBUG version."
-    perl -i -pe "s/'{{DEBUG_MODE}}'/True/g" "$utilsPath"
-else
-    echo "Building RELEASE version."
-    perl -i -pe "s/'{{DEBUG_MODE}}'/False/g" "$utilsPath"
-fi
+perl -i -pe "s/'{{DEBUG_MODE}}'/$debug/g" "$mainPath"
 
 python2 -m compileall ./build
+find ./build -name "*.py" -type f -exec rm {} \;
 
-cd ./as3
-rm ./bin/*.swf
-./build.sh
 
-cd ../
+
+# BUILD AS3
+if [ "$pythonOnly" = False ]; then
+  cd ./mod/as3
+  rm ./bin/*.swf
+  ./build.sh
+  cd ../../
+fi  
 
 mkdir -p ./build/res/gui/flash
-find ./as3/bin -name "*.swf" -exec cp {} ./build/res/gui/flash/ \;
+find ./mod/as3/bin -name "*.swf" -exec cp {} ./build/res/gui/flash/ \;
+
+
+
+# BUILD SERVER
+if [ "$modOnly" = False ] && [ "$pythonOnly" = False ] || [ ! -f ./cef-server/wotstat.widget.cef.zip ]; then
+  cd ./cef-server
+  ./build.sh
+  cd ../
+fi
+
+
 
 meta=$(<meta.xml)
 meta="${meta/\{\{VERSION\}\}/$v}"
-
+echo "$meta" > ./build/meta.xml
 cd ./build
-echo "$meta" > ./meta.xml
 
-folder="wotstat.cef_$v.wotmod"
-
-rm -rf $folder
 
 zip -dvr -0 -X $folder res -i "*.pyc"
 zip -dvr -0 -X $folder res -i "*.swf"
 zip -vr -0 -X $folder meta.xml
 
-cp ../wotstat.widget.cef.zip res/wotstat.widget.cef.zip
+cp ../cef-server/wotstat.widget.cef.zip res/wotstat.widget.cef.zip
 zip -dvr -0 -X $folder res/wotstat.widget.cef.zip
-
-# cp -r ../wotstat.widget.cef wotstat.widget.cef
-# zip -r res/wotstat.widget.cef.zip wotstat.widget.cef
-# rm -rf wotstat.widget.cef
-
-# zip -dvr -0 -X $folder res/wotstat.widget.cef.zip
 
 cd ../
 cp ./build/$folder $folder
